requirements.txt
# Web framework
flask
flask_sqlalchemy
flask_cors
werkzeug

# Database + vector search
psycopg2-binary
pgvector

# Embeddings (CPU-only)
sentence-transformers

# LLM inference (llama.cpp)
llama-cpp-python

# PDF + text utils
pdfminer.six
markdown2

# Environment variables
python-dotenv

# Production server
gunicorn

# LangChain utilities 
langchain
faiss-cpu

# Data processing
pandas
numpy

# Testing
pytest
pytest-mock
pytest-cov
pytest-benchmark
requests
deepeval
ragas